move interfaces into interface package
