+++ src/main/java/com/EighthLight/app/EasyDifficulty.java
@@ -1,5 +1,8 @@
 package com.EighthLight.app;
 
+import com.EighthLight.app.Interfaces.IBoard;
+import com.EighthLight.app.Interfaces.IStrategy;
+
 public class EasyDifficulty implements IStrategy {
     private String aiSymbol;
