+++ src/main/java/com/EighthLight/app/HardDifficulty.java
@@ -1,10 +1,13 @@
 package com.EighthLight.app;
 
+import com.EighthLight.app.Interfaces.IBoard;
+import com.EighthLight.app.Interfaces.IStrategy;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-public class HardDifficulty implements IStrategy{
+public class HardDifficulty implements IStrategy {
 
     private String aiSymbol;
     private String humanSymbol;
