--- src/main/java/com/EighthLight/app/IStrategy.java
@@ -1,5 +1,0 @@
-package com.EighthLight.app;
-
-public interface IStrategy {
-    void markBoard(IBoard board);
-}
