+++ src/main/java/com/EighthLight/app/Interfaces/IGame.java
@@ -1,4 +1,4 @@
-package com.EighthLight.app;
+package com.EighthLight.app.Interfaces;
 
 public interface IGame {
     void start();
