+++ src/main/java/com/EighthLight/app/Interfaces/IPlayer.java
@@ -1,5 +1,4 @@
-package com.EighthLight.app;
-
+package com.EighthLight.app.Interfaces;
 public interface IPlayer {
     String getSymbol();
     void makeMove(IBoard board);
