+++ src/main/java/com/EighthLight/app/Interfaces/IStrategy.java
@@ -1,0 +1,7 @@
+package com.EighthLight.app.Interfaces;
+
+import com.EighthLight.app.Interfaces.IBoard;
+
+public interface IStrategy {
+    void markBoard(IBoard board);
+}
