+++ src/main/java/com/EighthLight/app/MediumDifficulty.java
@@ -1,5 +1,8 @@
 package com.EighthLight.app;
 
+import com.EighthLight.app.Interfaces.IBoard;
+import com.EighthLight.app.Interfaces.IStrategy;
+
 public class MediumDifficulty implements IStrategy {
     public void markBoard(IBoard board) { }
 }
