+++ src/test/java/com/EighthLight/app/TestGame.java
@@ -1,5 +1,6 @@
 package com.EighthLight.app;
 
+import com.EighthLight.app.Interfaces.IPlayer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
